+++ NEWS.md
@@ -5,7 +5,11 @@
 * Regarding the spurious cases of the dilution curves: flag curves every time they have a PRA value close to 0.8?
 * Flag curves with negative correlation close to 1
 
+## Completed
+
+* Changed output excel sheet font to "Consolas" so that the number "0" and the letter "O" can be differentiated easily.
+
 # DCVtestkit 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.
-* Aim to create a tag version of this
+* Aim to create a tag version.